requirements.txt
Flask
opencv-python
dlib
imutils
python-vlc
matplotlib
